\name{c.simCT}
\alias{c.simCT}
\alias{print.simCT}
\title{
  Functions to Deal with Simulations of Camera Trap Studies
}
\description{
  The function \code{c} combine several sets of simulations generated by
  \code{simulateCTStudy5years}.  The function \code{print} print a
  description of objects returned by this function.
}
\usage{
\method{c}{simCT}(...)
\method{print}{simCT}(x,...)
}
\arguments{
  \item{\dots}{
    For \code{c}, objects of class \code{simCT} to be combined.  For
    \code{print}, additional objects to be passed from and to other
    functions. 
  }
  \item{x}{an object of class \code{simCT}}
}
\value{
  \code{c} returns an object of class \code{simCT}. \code{print} does
  not return anything (just prints a description of the object).
}
\references{

  Calenge et al. 2026. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.

}
\author{
  Clement Calenge \email{clement.calenge@ofb.gouv.fr}
}
\seealso{
  \code{\link{simulateCTStudy5years}}. 
}
\examples{

## Consider the dataset sim4iter (see the help page of this package
## to show how it was generated)
sim4iter

## Imagine a case where we had another similar object
anotherObject <- sim4iter ## just for the example

## combine sim4iter and anotherObject
(result <- c(sim4iter, anotherObject))

}
\keyword{ datagen }
