\name{dfPopSizeRedDeer}
\alias{dfPopSizeRedDeer}
\docType{data}
\title{
  Parameters to Simulate Changes in a Decreasing Red Deer Population in Chize 
}
\description{
  This dataset is a data.frame containing the parameters required to
  simulate the changes in a roe deer population over 5 years (20 percent
  decrease simulated).
}
\usage{data("dfPopSizeRedDeer")}
\format{
  A data frame with 5 observations (the 5 years) on the following 3 variables.
  \describe{
    \item{\code{Ntot}}{The total number of roe deer during a given year}
    \item{\code{Old}}{The number of roe deer surviving from the previous
      year.}
    \item{\code{New}}{The number of new roe deer born during the current
      year.}
  }
}
\source{
  Calenge et al. 2026. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\references{
  Calenge et al. 2026. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\examples{

dfPopSizeRedDeer

}
\keyword{datasets}
