\name{processCTStudy}
\alias{processCTStudy}
\alias{process_all}
\title{
  Summary of the Results of a Simulation of a Five-Year Camera Trap Study
}
\description{
  The function \code{processCTStudy} is used to calculate estimates of
  population size for each year, as well as the trend and rate of change
  between year 1 and year 5 (as well as 90\% confidence intervals and
  standard error for each parameter), using a chosen method.\cr
  The function \code{process_all} does the same thing for six different
  methods: (i)\code{REM_raw} is the random encounter model without accounting for
  imperfect detectability or different activity periods; (ii)
  \code{REM_detect} is the
  random encounter model accounting for imperfect detectability (the
  probability of encounter detection is estimated from the data) but
  without accounting for the presence of inactivity periods; (iii)
  \code{REM_detect_active} is a simila approach but accounting for both
  imperfect detectability and the presence of inactivity periods. The
  other three methods, \code{IS_raw}, \code{IS_detect},
  \code{IS_detect_active} are similar approaches, but using the instantaneous
  sampling rather than the importance sampling.\
}
\usage{

processCTStudy(listSim, surfaceChize=25979850, 
               method=c("REM","IS"),
               ideal=FALSE, detectability=TRUE,
               active=TRUE,
               stratified=FALSE, pathsBuffer=NULL,
               Nboot=1000, verbose=1,
               ibegin=1, iend=length(listSim$encounters))

process_all(listSim, stratified=FALSE, pathsBuffer=NULL)

}
\arguments{
  \item{listSim}{
    An object returned by the function \code{simulateCTStudy5years}.
  }
  \item{surfaceChize}{
    Numeric value indicating the surface area of the Chize study area
  }
  \item{method}{
    Character string. Either \code{"IS"} or \code{"REM"}.
  }
  \item{ideal}{
    Logical value.  Whether we simulate the case where all encounters
    have been detected (i.e., we set \code{detection = TRUE} for all
    encounters of the object).  This setting is only available for
    \code{method == "REM"}.
  }
  \item{detectability}{
    Logical value. Whether to account for imperfect detectability in the
    estimation
  }
  \item{active}{
    Logical value. Whether to remove detections occurring between 
    8AM and 5PM (when the animals are less active).
  }
  \item{Nboot}{
    Number of bootstrap samples
  }
  \item{stratified}{
    Logical value indicating whether the sample of camera traps was
    stratified or not (i.e. a given fraction of the traps was placed
    near paths).
  }
  \item{pathsBuffer}{
    if \code{stratified} is \code{TRUE}, an object of class \code{sf} or
    \code{sfc} containing the map of the 
    paths habitat used for stratification.
  }
  \item{verbose}{
    Integer value. Whether the function should provide the user with
    information. If 1, information on the simulation number is
    returned during the computation. In greater than 1, the information
    on the simulation number is provided as well as the information on
    lower-level processes (bootstrap, calculation of mean detectability).
  }
  \item{ibegin}{
    Integer value. The index of the first simulation to be processed
    (this calculation can be slow, so that the user may prefer to process
    only part of the simulations in \code{listSim}).
  }
  \item{iend}{
    Integer value. The index of the last simulation to be processed
    (this calculation can be slow, so that the user may prefer to process
    only part of the simulations in \code{listSim}).
  }
  
}
\value{
  The function \code{processCTStudy} returns a list with one element per simulation.  This
  element is a data.frame containing the following columns: (i)
  \code{PtEst}: the point estimate of population sizes, trend and rate
  of change, (ii) \code{EstSE}: the standard error of the bootstrap samples, (iii) \code{EstCIl} and
  \code{estCIu}: the lower and upper limit of the 90\% confidence
  interval on the parameter (rownames indicate the parameters).\cr

  The function \code{process_all} returns a list with one component per
  method, each component corresponding to an object similar to the one
  returned by \code{processCTStudy}
  
}
\references{

  Calenge et al. 2025. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.

}
\author{
  Clement Calenge \email{clement.calenge@ofb.gouv.fr}
}
\seealso{
  \code{\link{estimIS}} for more details on the population size
  estimation using instantaneous sampling and random encounter models. 
}
\examples{

## Consider the example dataset sim4iter generated by simulateCTStudy5years
## Use the random encounter model, correcting for detectability and
## focusing only on active periods. 
processCTStudy(sim4iter, method="REM", Nboot=200)



## Instantaneous sampling and supposing a known detectability
## and using all detections, including during inactive periods
processCTStudy(sim4iter, method="IS", Nboot=200,
               detectability=FALSE, active=FALSE)


## All methods
\dontrun{
process_all(sim4iter)
}




}
\keyword{robust }
\keyword{distribution }
