\name{removeHour}
\alias{removeHour}
\title{
  Remove all Detections occurring Between Two Times During the Day
}
\description{
  \code{removeHour} removes all the detections that occurred between two
  times during the day (typically between 8:00 and 17:00
}
\usage{
removeHour(x, startHour=8, endHour=17)
}
\arguments{
  \item{x}{

    A data.frame containing the results of a camera trap monitoring
    during 30 days of a given year.  This data.frame contain one row per
    encounter between an animal and a camera trap during the study. It
    must contain the following variables: (i) \code{detection} is a
    logical value indicating if the encounter has been detected by the
    camera trapreturned; (ii) \code{whichani} is an integer variable
    indicating the animal ID for the encounter, (iii) \code{whichCT} is
    an integer variable indicating the camera trap ID for the encounter,
    (iv) \code{beginning}: an integer value indicating the date/time of
    the beginning of the encounter (converted as integer from the class
    \code{POSIXct}), (v) \code{beginning}: an integer value indicating
    the date/time of the end of the encounter (converted as integer from
    the class \code{POSIXct}), (vi) \code{duration}: the duration of the
    encounter in seconds, (vii) \code{Pdetect}: the theoretical
    probability of detection (as estimated by a survival model), (viii)
    \code{when}: a list indicating, for each encounter, the date/time of
    all associations during the encounter, expressed as integer values
    (converted from the class \code{POSIXct}.

  }
  \item{startHour}{
    Integer value indicating the starting time of the "inactive" period.
  }
  \item{endHour}{
    Integer value indicating the ending time of the "inactive" period.

  }
}
\value{
  An object similar to \code{x}, but with relevant detection removed.
}
\references{
  Calenge et al. 2025. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\author{
  Clement Calenge \email{clement.calenge@ofb.gouv.fr}
}
\seealso{
  \code{\link{estimIS}}
}
\examples{
## Consider the example dataset sim4iter generated by simulateCTStudy5years
## Focus on the first iteration (= first study), and the first year
firstyear <- sim4iter$encounters[[1]]$results[[1]]

## As a data.frame
firstyeardf <- do.call(rbind, firstyear)

## Remove active periods
firstyeardfactive <- removeHour(firstyeardf)

## Less detection in the latter
dim(firstyeardf)
dim(firstyeardfactive)

## and the times:

## Not filtered
timeNotFiltered <- unlist(firstyeardf$When)
class(timeNotFiltered) <- c("POSIXct","POSIXt")
## Number of detections during each hour
table(as.POSIXlt(timeNotFiltered)$hour)

## Focus on active periods
timeFiltered <- unlist(firstyeardfactive$When)
class(timeFiltered) <- c("POSIXct","POSIXt")
## Number of detections during each hour
table(as.POSIXlt(timeFiltered)$hour)

## Note that internally, by convention, we have chosen to make the day
## start at 17h. So that 8AM (real time) corresponds to 15:00
## (transformed time). So that no detection occurs after 15:00 in the
## transformed dataset.

}
\keyword{models }
