\name{roeDeerMovement}
\alias{roeDeerMovement}
\title{
  Simulate the Movement of a Roe Deer
}
\description{
  The function \code{roeDeerMovement} simulates the movement of a roe
  deer, including inactive periods (with nycthemeral rhythm), intensive
  use of food patches, and between-patches movements.
}
\usage{
roeDeerMovement(nj, lixy, verbose = TRUE, complete = FALSE, lipt = NULL)
}
\arguments{
  \item{nj}{
    The number of days during which the movement should be simulated
  }
  \item{lixy}{
    A list of set of patches. Each component of the list is a matrix
    with two columns containing the coordinates of patches within the
    home range of a roe deer. 
  }
  \item{verbose}{
    Logical value indicating whether information should be returned to
    the user.
  }
  \item{complete}{
    Logical value indicating which information should be returned to the
    user. By default, only the xy coordinates and date/time of simulated
    relocations are returned. If \code{TRUE}, additional information is
    returned (see below).
  }
  \item{lipt}{
    
    A list with two components named \code{ptg} and \code{ptp}
    corresponding to the set of matrices to be used to simulate
    Orntein-Uhlenbeck with Foraging processes for the between-patches
    and within-patches movement respectively (each component is
    generated by the function \code{prepareMvt}, see the help page of
    this function).
  
}
}
\details{
  This function simulates the movement of a roe deer, with one
  relocation per second for \code{nj} days, accounting for the
  nycthemeral rhythm. 
}
\value{
  A data.frame containing the x and y coordinates of the relocations, as
  well as the corresponding date/time.  If \code{complete} is
  \code{TRUE}, the data.frame also contains whether the animal was
  active or not, which patch attracted the movement for all relocations,
  and the burst ID (which could be used to separate the different
  components of the movement -- food patches, between patches movement,
  inactive burst, etc.). 
}
\references{
  Calenge et al. 2026. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\author{
  Clement Calenge \email{clement.calenge@ofb.gouv.fr}
}
\seealso{
  \code{\link{prepareMvt}} to prepare the matrices used to simulate the
  OUF processes.
}
\examples{

## We use this list of set of patches:
str(listPatches)

## Let's say that we will simulate animals with the first
## set of patches:
setpat <- list(listPatches[[1]])

## We use the dataset mvtMatrices for the set of matrices
str(mvtMatrices)

## We will simulate an animal with sigma=1000 for between-patches
## movements (first matrix of listMatBetween).
lipt <- list(ptg=mvtMatrices$listMatBetween[[1]],
             ptp=mvtMatrices$MatWithin)

## Simulation of movement for 30 days:
set.seed(777) ## for reproducibility
rd <- roeDeerMovement(30, setpat, lipt = lipt)

## movement (centroid = c(0,0))
plot(rd[,1:2], ty="l")


}
\keyword{datagen}
