\name{sim4iter}
\alias{sim4iter}
\docType{data}
\title{
  Example of Simulation of a Camera Trap Study
}
\description{
  This dataset has been generated by the function
  \code{simulateCTStudy5years} (4 iterations).
}
\usage{data("sim4iter")}
\format{
  This object a list returned by the function
  \code{simulateCTStudy5years} (see the help page of this function, for
  more details on this format).
}
\details{
  See the examples for a description of the code that generated this
  file.
}
\source{
  Calenge et al. 2025. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\references{
  Calenge et al. 2025. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\examples{

###########################################
##
## WARNING !!!! THE FOLLOWING CODE CAN BE VERY SLOW
## It took 41 minutes to execute on an intel core I7.
## The resulting dataset is available as a dataset
## 
\dontrun{

set.seed(777) ## for reproducibility
    
sim4iter <- simulateCTStudy5years(nCT = 100, duration = 30,
                                  niter = 4, dfPopSize,
                                  listPatches, contourChize, contourN=northChize,
                                  contourS=southChize,
                                  dff=relationAnimalSigmaMaxd, habitatMap,
                                  listptg=mvtMatrices$listMatBetween,
                                  ptp=mvtMatrices$MatWithin, nofi = 0, backup = TRUE)

}



## The resulting object:
str(sim4iter,1)

## Focus on the main component "encounter"
str(sim4iter$encounter,2)

## Focus on the first iteration and on the component results
str(sim4iter$encounter[[1]]$results,1)

## One component per year. Focus on -- say -- the second year
str(sim4iter$encounter[[1]]$results[[2]],1)

## Each data.frame describes all the encounters between a pair of animal/CT
## Example:
sim4iter$encounter[[1]]$results[[2]][[1]]

}
\keyword{datasets}
