\name{simCTChize-package}
\alias{simCTChize-package}
\alias{simCTChize}
\docType{package}
\title{
\packageTitle{simCTChize}
}
\description{
\packageDescription{simCTChize}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simCTChize}
\packageIndices{simCTChize}
}
\author{
\packageAuthor{simCTChize}

Maintainer: \packageMaintainer{simCTChize}
}
\references{
 References.
}
\keyword{ package }
\seealso{
  \code{\link{simulateCTStudy5years}}
}
\examples{
## No examples
}
