\name{simulateCTStudy5years}
\alias{simulateCTStudy5years}
\title{
  Simulate a Camera Trap Study of a Roe Deer Population During 5 years
}
\description{
  Simulate a Camera Trap Study of a Roe Deer Population During 5 years
}
\usage{
simulateCTStudy5years(nCT = 100, duration = 30, niter = 15, dfPopSize,
                      listPatches, contourChize, contourN, contourS,
                      dff, habitatMap, listptg, ptp,
                      stratSampling=FALSE, propPaths=NULL, pathsMap=NULL,
                      pathsBuffer=NULL, offPathsMap=NULL, offPathsBuffer=NULL,
                      habitatSelection=FALSE, LRroeDeer=FALSE,
                      LRroeDeerIncreaseDV=FALSE,
                      fieldangle=0.175, depth=20,
                      nofi = 0, backup = TRUE)
}
\arguments{
  \item{nCT}{
    The number of camera traps to simulate on the study area
  }
  \item{duration}{
    The number of days of monitoring
  }
  \item{niter}{
    The number of simulations to carry out. WARNING, one iteration
    can take several hours.
  }
  \item{dfPopSize}{
    A data.frame with three columns named Ntot, Old and New containing
    respectively the total number of roe deer during each one of the
    five years, the number of roe deer surviving from the previous years
    and the number of roe deer born each year. The dataset
    \code{dfPopSize} is the one used in the paper.
  }
  \item{listPatches}{
    A list of set of patch centroid, with the same structure as the
    dataset \code{listPatches} from the package (only used if
    \code{habitatSelection} is \code{FALSE}).
  }
  \item{contourChize}{
    A two-columns matrix containing the xy coordinates of the limit of
    the Chize study area. Available in the dataset \code{contourChize}.
  }
  \item{contourN}{
    A two-columns matrix containing the xy coordinates of the limit of
    the northern part of the Chize study area. Available in the dataset
    \code{northChize}.     
  }
  \item{contourS}{
    A two-columns matrix containing the xy coordinates of the limit of
    the northern part of the Chize study area. Available in the dataset
    \code{southChize}. 
  }
  \item{dff}{    
    A data.frame containing the maximum predicted distance between the
    animal movement and its home-range centroid (stored in a column
    named \code{dmaxp}) as a function of the animal number (stored in
    the column \code{ani}; each number corresponding to a set of patch
    in \code{listPatches}) and of a value of \code{sigma} (stored in the
    column \code{sigma}, where the actual value of sigma used for the
    OUf process is this value times 1000). See the dataset
    \code{relationAnimalSigmaMaxd} for such a dataset.
  }
  \item{habitatMap}{
    An object of class \code{sf} (from the package of the same name)
    storing the habitat type in Chize (in a column named
    \code{habitat}).  Three habitat types are 
    possible: \code{"Coppice"}, \code{"REG"} (for regeneration) and
    \code{"OPE"} (for open). See the dataset \code{habitatMap} for such
    a map of Chize.
  }
  \item{listptg}{
    A list of 10 sets of 3 matrices used to simulate the between-patches
    OUF process. Each component is generated by the function
    \code{prepareMvt} (see the help page of this function), and should
    correspond to a value of \code{sigma} (comprised between 1000 and
    10000). See examples. 
  }
  \item{ptp}{
    A list of 3 matrices used to simulate the within-patches
    OUF process. Each component is generated by the function
    \code{prepareMvt} (see the help page of this function).
  }
  \item{stratSampling}{
    Logical value.  Whether (\code{TRUE}) or not (\code{FALSE}) the
    camera traps sampling should be stratified (with a given proportion
    of traps located close to paths).
  }
  \item{propPaths}{
    Numeric value. If the sampling is stratified, the proportion of
    camera traps placed close to paths.
  }
  \item{pathsMap}{
    An object of class \code{sf/sfc} (from the package of the same name)
    storing the location of the paths in the study area (polygon
    features).  See the dataset \code{pathsMap} for such a map in 
    Chize (it is actually a 50 cm buffer around the paths linear
    features).
  }
  \item{pathsBuffer}{
    An object of class \code{sf/sfc} (from the package of the same name)
    storing the map of all areas located at less than 20 m from a path
    in the study area (polygon features).  This distance of 20 m is
    necessary as this is the maximum radius of the detection zone of a
    camera trap.  This map is used to identify all areas detectable from
    a trap located on the path See the dataset \code{pathsBuffer}
    for such a map in Chize. 
  }
  \item{offPathsMap}{
    An object of class \code{sf/sfc} (from the package of the same name)
    storing the map of all areas located at more than 20 m from a path
    in the study area (polygon features).  See the dataset
    \code{offPathsMap} for such a map in Chize. 
  }
  \item{offPathsBuffer}{
    An object of class \code{sf/sfc} (from the package of the same name)
    storing the map of all areas located at more than 40 m from a path
    in the study area (polygons).  This map is used to make that the
    detection zone of all sampled CTs in the off-paths area does not
    intersect the area close to the paths (which is selected by the
    deer). See the dataset \code{offPathsBuffer} for such a map in Chize. 
  }
  \item{habitatSelection}{
    Logical value.  Whether (\code{TRUE}) or not (\code{FALSE}) a
    habitat selection by the roe deer should be simulated. In this case,
    a centroid is randomly sampled on the north or southern par of the
    study area (with respective probability 193/300 and 1-193/300). Then
    a random number of patches (uniform between 3 and 10) is selected in
    the areas located at less than 20 m from the paths within a random
    distance from this centroid (this distance is drawn randomly from a
    uniform distribution bounded by 100 m and 300 m). These patches are
    used as center of attractions in the function
    \code{roeDeerMovement}.  See the function \code{HSlocation}.
  }
  \item{LRroeDeer}{
    
    Logical value. If \code{TRUE}, the function simulates much larger
    home ranges than for the standard roe deer.  More precisely, the centroid of
    the home range is randomly drawn in the "core study area" (eroded
    study area corresponding to an inner buffer of 1000 m from the limit
    of the study area), possibly restricted on paths if \code{habitat
      selection} is \code{TRUE}. Then a random number of patches
    comprised between 5 and 10 is randomly drawn in a box of +/- 2000 m
    centred on this centroid (possibly only on paths if
    \code{habitatSelection} is TRUE). Then, large patches centred on
    these attraction points are simulated using mixed Ornstein-Uhlenbeck
    movement characterized by a large value of sigma (equal to 10000,
    see the function \code{prepareMvt}). 

  }
  \item{LRroeDeerIncreaseDV}{
    
    Logical value. If \code{TRUE}, the function simulates an increasing 
    average home range over the five years: the patches simulated are
    randomly drawn in a box of +/- X m centred on this centroid, where X
    increases from 1414 m to 
    2000 m over the 5 years. Then, large patches centred on
    these attraction points are simulated using mixed Ornstein-Uhlenbeck
    movement characterized by a large value of sigma (equal to 10000,
    see the function \code{prepareMvt}). 

  }
  \item{fieldangle}{
    The angular width of the detection zone (in radians)
  }
  \item{depth}{
    The radius (in metres) of the detection zone.
  }
  \item{nofi}{
    Integer. A value to be used as a seed for the random number
    generator in \code{set.seed}.
  }
  \item{backup}{
    Logical value indicating whether the simulations should be saved in
    Rds files on the disk (strongly recommended).
  }
}
\details{

  This function first randomly places \code{nCT} camera traps over the
  study area of Chize (and defines a probability of detection of animals
  depending on the habitat type, see the function \code{organiseCT} for
  more details). Then it simulates a number of roe deer (or large-range
  roe deer) for each year given in \code{dfPopSize}. Their movements are
  simulated with the function \code{roeDeerMovement}. Their home-range
  size increase as the population decrease. The function then counts the
  number of detected encounters and associations between each animal and
  each camera trap.
  
}
\value{

  This function is so long (can take more than a day for even a moderate
  number of iterations) that the user will probably be more interested
  in the backup files generated. Three RDS files will be generated: (i)
  A file with a name of the type
  "fic5years-\code{nofi}listSimus-CT\code{nCT}-Ani.Rds" containing the
  main results (see below), (ii) A file with a name of the type
  "fic5years-\code{nofi}listSpeed-CT\code{nCT}-Ani.Rds" containing for
  each iteration a list with \code{niter} components, each component
  being itself a list with 5 subcomponents (the 5 years), each
  subcomponent being a vector of the average speed of each simulated
  animal (meters/second), and (iii) a file with a name of the type
  "fic5years-\code{nofi}listActiveSpeed-CT\code{nCT}-Ani.Rds", similar
  to the previous one, but focusing only on active periods. Note that
  the function itself returns a list with three components corresponding
  to these three files.

  The file containing the main results is a list with \code{niter}
  components, each component being a list with three subcomponent: (i) a
  vector named \code{creationCT} with two components (the seed used by
  the random number generator to generate the position of the camera
  traps on the area, and the seed used by this generator to generate the
  position of trees -- obstacles -- in the detection zone), (ii) a
  data.frame named \code{indicetraps} containing the coordinates x and y
  of the camera traps, as well as the habitat type in which they fall,
  (iii) a list named \code{results} containing the encounter and
  association detections. Each component of this list corresponds to one
  year and is itself a list containing the set of encounters between a
  given animal and a given camera trap. These encounters are stored as
  data.frames with one row per encounter between an animal and a camera
  trap during the study. It contains the following variables: (i)
  \code{detection} is a logical value indicating if the encounter has
  been detected by the camera trapreturned; (ii) \code{whichani} is an
  integer variable indicating the animal ID for the encounter, (iii)
  \code{whichCT} is an integer variable indicating the camera trap ID
  for the encounter, (iv) \code{beginning}: an integer value indicating
  the date/time of the beginning of the encounter (converted as integer
  from the class \code{POSIXct}), (v) \code{beginning}: an integer value
  indicating the date/time of the end of the encounter (converted as
  integer from the class \code{POSIXct}), (vi) \code{duration}: the
  duration of the encounter in seconds, (vii) \code{Pdetect}: the
  theoretical probability of detection (as estimated by a survival
  model), (viii) \code{when}: a list indicating, for each encounter, the
  date/time of all associations during the encounter, expressed as
  integer values (converted from the class \code{POSIXct}.

}
\references{
  Calenge et al. 2026. Evaluating camera trap methods for monitoring
  population trends in ungulates: insights from simulation. in prep.
}
\author{
  Clement Calenge \email{clement.calenge@ofb.gouv.fr}
}
\seealso{
  \code{\link{organiseCT}} for more information on how the camera traps
  are simulated, and \code{\link{roeDeerMovement}} for more information
  on how the camera traps are simulated.
}
\examples{

###########################################
##
## WARNING !!!! THE FOLLOWING CODE CAN BE VERY SLOW
## It took 41 minutes to execute on an intel core I7.
## The resulting dataset is available as a dataset (see below)
## 
\dontrun{

## For reproducibility
set.seed(777)
    
sim4iter <- simulateCTStudy5years(nCT = 100, duration = 30,
                                  niter = 4, dfPopSize,
                                  listPatches, contourChize, contourN=northChize,
                                  contourS=southChize,
                                  dff=relationAnimalSigmaMaxd, habitatMap,
                                  listptg=mvtMatrices$listMatBetween,
                                  ptp=mvtMatrices$MatWithin, nofi = 0, backup = TRUE)

}

## The resulting object:
str(sim4iter,1)

## Focus on the main component "encounter"
str(sim4iter$encounter,2)

## Focus on the first iteration and on the component results
str(sim4iter$encounter[[1]]$results,1)

## One component per year. Focus on -- say -- the second year
str(sim4iter$encounter[[1]]$results[[2]],1)

## Each data.frame describes all the encounters between a pair of animal/CT
## Example:
sim4iter$encounter[[1]]$results[[2]][[1]]

}
\keyword{ datagen }
